+++ modulus/distributed/mappings.py
@@ -22,52 +22,53 @@ class _CopyToParallelRegion(torch.autograd.Function):
     """Pass the input to the parallel region"""
 
     @staticmethod
-    def symbolic(graph, input_, group_):
+    def symbolic(graph, input_, group_):  # pragma: no cover
         return input_
 
     @staticmethod
-    def forward(ctx, input_, group_):
+    def forward(ctx, input_, group_):  # pragma: no cover
         ctx.group = group_
         return input_
 
     @staticmethod
-    def backward(ctx, grad_output):
-        return _reduce(grad_output, group=DistributedManager().group(ctx.group))
+    def backward(ctx, grad_output):  # pragma: no cover
+        return _reduce(grad_output, group=DistributedManager().group(ctx.group)), None
 
 
 class _ReduceFromParallelRegion(torch.autograd.Function):
     """All-reduce the input from the parallel region"""
 
     @staticmethod
-    def symbolic(graph, input_, group_):
+    def symbolic(graph, input_, group_):  # pragma: no cover
         return _reduce(input_, group=DistributedManager().group(group_))
 
     @staticmethod
-    def forward(ctx, input_, group_):
+    def forward(ctx, input_, group_):  # pragma: no cover
         return _reduce(input_, group=DistributedManager().group(group_))
 
     @staticmethod
-    def backward(ctx, grad_output):
-        return grad_output
+    def backward(ctx, grad_output):  # pragma: no cover
+        return grad_output, None
 
 
 class _ScatterToParallelRegion(torch.autograd.Function):
     """Split the input and keep only the chunk corresponding to the rank."""
 
     @staticmethod
-    def symbolic(graph, input_, dim_, group_):
+    def symbolic(graph, input_, dim_, group_):  # pragma: no cover
         return _split(input_, dim_, group=DistributedManager().group(group_))
 
     @staticmethod
-    def forward(ctx, input_, dim_, group_):
+    def forward(ctx, input_, dim_, group_):  # pragma: no cover
         ctx.dim = dim_
         ctx.group = group_
         return _split(input_, dim_, group=DistributedManager().group(group_))
 
     @staticmethod
-    def backward(ctx, grad_output):
+    def backward(ctx, grad_output):  # pragma: no cover
         return (
             _gather(grad_output, ctx.dim, group=DistributedManager().group(ctx.group_)),
+            None,
             None,
         )
 
@@ -76,19 +77,20 @@ class _GatherFromParallelRegion(torch.autograd.Function):
     """Gather the input from parallel region and concatenate."""
 
     @staticmethod
-    def symbolic(graph, input_, dim_, group_):
+    def symbolic(graph, input_, dim_, group_):  # pragma: no cover
         return _gather(input_, dim_, group=DistributedManager().group(group_))
 
     @staticmethod
-    def forward(ctx, input_, dim_, group_):
+    def forward(ctx, input_, dim_, group_):  # pragma: no cover
         ctx.dim = dim_
         ctx.group = group_
         return _gather(input_, dim_, group=DistributedManager().group(group_))
 
     @staticmethod
-    def backward(ctx, grad_output):
+    def backward(ctx, grad_output):  # pragma: no cover
         return (
             _split(grad_output, ctx.dim, group=DistributedManager().group(ctx.group)),
+            None,
             None,
         )
 
@@ -102,20 +104,21 @@ class _GatherWithinParallelRegion(torch.autograd.Function):
     """
 
     @staticmethod
-    def symbolic(graph, input_, dim_, group_):
+    def symbolic(graph, input_, dim_, group_):  # pragma: no cover
         return _gather(input_, dim_, group=DistributedManager().group(group_))
 
     @staticmethod
-    def forward(ctx, input_, dim_, group_):
+    def forward(ctx, input_, dim_, group_):  # pragma: no cover
         ctx.dim = dim_
         ctx.group = group_
         return _gather(input_, dim_, group=DistributedManager().group(group_))
 
     @staticmethod
-    def backward(ctx, grad_output):
+    def backward(ctx, grad_output):  # pragma: no cover
         red = _reduce(grad_output, group=DistributedManager().group(ctx.group_))
         return (
             _split(red, ctx.dim, group=DistributedManager().group(ctx.group)),
+            None,
             None,
         )
 
@@ -123,27 +126,27 @@ class _GatherWithinParallelRegion(torch.autograd.Function):
 # -----------------
 # Helper functions.
 # -----------------
-def copy_to_parallel_region(input, group):
+def copy_to_parallel_region(input, group):  # pragma: no cover
     """Copy input"""
     return _CopyToParallelRegion.apply(input, group)
 
 
-def reduce_from_parallel_region(input, group):
+def reduce_from_parallel_region(input, group):  # pragma: no cover
     """All-reduce the input from the matmul parallel region."""
     return _ReduceFromParallelRegion.apply(input, group)
 
 
-def scatter_to_parallel_region(input, dim, group):
+def scatter_to_parallel_region(input, dim, group):  # pragma: no cover
     """Split the input and keep only the corresponding chuck to the rank."""
     return _ScatterToParallelRegion.apply(input, dim, group)
 
 
-def gather_from_parallel_region(input, dim, group):
+def gather_from_parallel_region(input, dim, group):  # pragma: no cover
     """Gather the input from matmul parallel region and concatenate."""
     return _GatherFromParallelRegion.apply(input, dim, group)
 
 
-def gather_within_parallel_region(input, dim, group):
+def gather_within_parallel_region(input, dim, group):  # pragma: no cover
     """
     Gather the input within parallel region and concatenate.
     The same forward method as gather_from_parallel_region, the difference is only in