+++ CHANGELOG.md
@@ -11,12 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added Stokes flow dataset
-- An experimental version of SFNO to be used in unified training recipe for weather models
+- An experimental version of SFNO to be used in unified training recipe for
+weather models
 - Added distributed FFT utility.
 - Added ruff as a linting tool.
 - Ported utilities from Modulus Launch to main package.
 - EDM diffusion models and recipes for training and sampling.
 - NGC model registry download integration into package/filesystem.
+- Added distributed process group configuration mechanism.
+- Added DistributedManager utility to instantiate process groups based on they
+process group config
 
 ### Changed